Update services.xml
Add `<search/>` to the feed container nodes.
Otherwise the index requests will end up in your query nodes. 
Not sure on why this happens. I just noticed that if I add `<search>` to the feed container then the cluster behaves as I expect, and does not use the query nodes while indexing.